Conditional build test tags to fix test build on Windows
